requirements.txt
altgraph==0.17.2
autopep8==1.6.0
future==0.18.2
pefile==2021.9.3
pycodestyle==2.8.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
python-vlc==3.0.16120
pywin32-ctypes==0.2.0
toml==0.10.2
